requirement.txt
asgiref==3.8.1
contourpy==1.2.1
crispy-bootstrap5==2024.2
cycler==0.12.1
Django==4.2
django-admin-model-list-order==0.0.4
django-appconf==1.0.6
django-autocomplete-light==3.11.0
django-awesome-avatar==1.1.3
django-clean==0.1.2
django-crispy-forms==2.2
django-extra-views==0.14.0
django-formtools==2.5.1
django-grappelli==4.0.1
django-htmx==1.17.2
django-imagekit==5.0.0
django-model-utils==4.5.0
django-phonenumber-field==8.0.0
django-select2==8.2.1
django-session-timeout==0.1.0
django-static-fontawesome==6.4.2.1
django-utils2==3.0.2
django-view-breadcrumbs==2.5.1
djhtml==3.0.6
fonttools==4.53.1
Image-Kit==0.0.3
kiwisolver==1.4.5
matplotlib==3.9.1
numpy==2.0.0
packaging==24.1
phonenumbers==8.13.40
pilkit==3.0
pillow==10.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-utils==3.8.2
scipy==1.14.0
six==1.16.0
sqlparse==0.5.0
sweetify==2.3.1
typing_extensions==4.12.2
